✨ Revolutionize the site layout ✨
Too many changes to list all in detail, lol

The new site, which is built using nested flexboxes has 2 sections
- A welcome section on the left, which contains a vertically centered bio (coming soon) and the clock on the bottom
- A widgets section on the right which contains the good ol' search box, the answer box (now renamed content, since the plan is to add more stuff that can't be classified as an answer to the query) and a list of other projects (coming soon)

I also changed the text sizing almost everywhere and added box shadow. As for the colors I reduced the background color's contrast to be better on the eyes and added a background color to the search box.