Part 1 of implementing commands of type widget
Now whenever the user types a command that's of type widget a child element of with tag name 'COMMANDNAME-widget' is created under the widget element, and the properties the user has typed are added. After the user has typed the first word, only the properties change - the entire widget isn't needlessly recreated.