Implement GIF widget
Implemented GIF widget first, as it is of utmost importance to be able to fetch and display funny GIFs from the internet.

Searching for "gif $gif_content" (e.g. gif cat) displays a centerd image of a GIF with a title, both of which are fetched from tenor as in the original version of the startpage. An async lit js task was used in combination with an AbortController and a timeout to throttle requests (i.e. wait 300ms before making a request and cancel it if the user input changes).

PS: The word "$object" was replaced with the more specific and clear "$gif_content"