Improve command previewing & add comments
When typing a command that corresponds to a website, the action to take when enter is pressed changes based on the parameters the user has typed. For instance, show "search for cat on google.com in a new tab" instead of "search for $query on google.com in a new tab" when the user enters the input "g cat".

The different cases the "previewCommand" function deals with are also changed. Don't show the command info when input that starts with a space is entered, by checking that the inputWords' length is one. Additionally, show "invalid or no command type" if the command type in the commands.json is missing / invalid.

To achieve all this and also make the code simpler and more readable, the "previewCommand" function is run *every* time the user inputs something, and takes on the functionality that was in the input event listener.

Finally add some (hopefully) helpful comments in the "previewCommand" and "executeCommand".